+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="AIzaSyCa3x16MPeqXAaf35RSN8dFd0rBRJwybNo" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.StudyBuddies">
